# lib/api-zod/src/generated/types/budgetInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ExpenseFlow finance API
 * OpenAPI spec version: 0.2.0
 */
import type { TransactionCategory } from './transactionCategory';

export interface BudgetInput {
  category: TransactionCategory;
  /** @exclusiveMinimum 0 */
  amount: number;
  /**
     * @minimum 1
     * @maximum 12
     */
  month: number;
  /** @minimum 2000 */
  year: number;
}