# lib/api-zod/src/generated/types/budgetUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ExpenseFlow finance API
 * OpenAPI spec version: 0.2.0
 */

export interface BudgetUpdate {
  /** @exclusiveMinimum 0 */
  amount?: number;
}