# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ExpenseFlow finance API
 * OpenAPI spec version: 0.2.0
 */

export interface DashboardSummary {
  /** @pattern ^[0-9]{4}-[0-9]{2}$ */
  month: string;
  balance: number;
  totalIncome: number;
  totalExpenses: number;
  monthlyExpenses: number;
  transactionCount: number;
  budget: number;
  budgetUsedPercent: number;
}