# lib/api-zod/src/generated/types/incomeExpensePoint.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ExpenseFlow finance API
 * OpenAPI spec version: 0.2.0
 */

export interface IncomeExpensePoint {
  /** @pattern ^[0-9]{4}-[0-9]{2}$ */
  month: string;
  label: string;
  income: number;
  expense: number;
}