# lib/api-zod/src/generated/types/paymentMethod.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ExpenseFlow finance API
 * OpenAPI spec version: 0.2.0
 */

export type PaymentMethod = typeof PaymentMethod[keyof typeof PaymentMethod];


export const PaymentMethod = {
  Card: 'Card',
  Cash: 'Cash',
  Bank_transfer: 'Bank transfer',
  UPI: 'UPI',
} as const;