# lib/api-zod/src/generated/types/transactionUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * ExpenseFlow finance API
 * OpenAPI spec version: 0.2.0
 */
import type { PaymentMethod } from './paymentMethod';
import type { TransactionCategory } from './transactionCategory';
import type { TransactionType } from './transactionType';

export interface TransactionUpdate {
  /** @exclusiveMinimum 0 */
  amount?: number;
  /** @minLength 1 */
  description?: string;
  merchant?: string;
  type?: TransactionType;
  category?: TransactionCategory;
  /** @pattern ^[0-9]{4}-[0-9]{2}-[0-9]{2}$ */
  date?: string;
  paymentMethod?: PaymentMethod;
  /** @nullable */
  notes?: string | null;
}